Move implementaion files to impl/ subdirectory
Modify the package declaration
Import necessary interface for each implementation
